+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes
 - Actually fail and give SYNTAX_ERROR when input cannot be entirely parsed
 - Add --max-infractions option to magik-lint
 - FileNotInLoadListCheck trims lines from load_list
+- Fix bug where FormattingCheck does not handle augmented assignments properly
 
 
 0.2.0 (2019-08-31)